pyproject.toml
[project]
name = "er04"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = []

[dependency-groups]
dev = [
    "adr>=0.4.0",
    "docling>=2.57.0",
    "genanki>=0.13.1",
    "llama-index-core>=0.14.5",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-llms-huggingface>=0.6.1",
    "llama-index-llms-huggingface-api>=0.6.1",
    "llama-index-node-parser-docling>=0.4.1",
    "llama-index-readers-docling>=0.4.1",
    "llama-index-readers-file>=0.5.4",
    "llama-index-vector-stores-milvus>=0.9.3",
    "pre-commit>=4.3.0",
    "pymilvus[milvus-lite]>=2.6.2",
    "python-dotenv>=1.1.1",
    "ruff>=0.14.0",
    "transformers>=4.57.1",
    "ty>=0.0.1a22",
]

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "v$version"
use_pre_commit = true
